admin_panel/movies_admin/requirements/dev.txt
-r base.txt
django-debug-toolbar==3.2.1
django-extensions==3.1.3
factory-boy==3.2.0
Faker==8.5.1
ipython==7.23.1
mixer==7.1.2
